moved axios functions
