+++ src/components/App/App.jsx
@@ -1,10 +1,9 @@
-
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
-import Header from '../Header/Header.jsx';
-import ItemForm from '../ItemForm/ItemForm.jsx';
-import ShoppingList from '../ShoppingList/ShoppingList.jsx';
-import './App.css';
+import React, { useState, useEffect } from "react";
+import axios from "axios";
+import Header from "../Header/Header.jsx";
+import ItemForm from "../ItemForm/ItemForm.jsx";
+import ShoppingList from "../ShoppingList/ShoppingList.jsx";
+import "./App.css";
 
 function App() {
   const [items, setItems] = useState([]);
@@ -14,32 +13,15 @@ function App() {
   }, []);
 
   const getItems = () => {
-    axios.get('/api/shopping')
+    axios
+      .get("/api/shopping")
       .then((response) => {
         setItems(response.data);
       })
       .catch((error) => {
-        console.log('Error with getItems()', error);
+        console.log("Error with getItems()", error);
       });
   };
-
-
-
-/*
-const deleteItem=(id) =>{
-    axios.delete('/shopping/${id}')
-    .then(()={})
-}
-*/
-
-
-//list item  put 
-//shopping
-
-
-
-
-
 
   return (
     <div className="App">
@@ -53,8 +35,3 @@ const deleteItem=(id) =>{
 }
 
 export default App;
-
-
-
-
-
